# New method for inserting and updating route