## Summary

This PR introduces three new public methods to the `Router` API: `Router.NewRoute`, `Router.HandleRoute`, and `Route.UpdateRoute`. These methods were previously part of Fox but were not exposed publicly. They allow for a more flexible and decoupled approach to route creation and configuration before registering them in the router.

## Changes Introduced

### New Public Methods

- **`Router.NewRoute(pattern string, handler http.Handler, opts ...Option) (*Route, error)`**
  - Instantiates a new route with a pattern, handler, and optional configurations.
- **`Router.HandleRoute(method string, route *Route) error`**
  - Registers an existing `Route` object for a given HTTP method.
- **`Router.UpdateRoute(method string, route *Route) error`**
  - Updates an existing route in the router with new configurations.

### Example Usage

#### Creating and Registering a Route

```go
f, _ := fox.New()

// Create a new route instance
route, _ := f.NewRoute("/bar", Handler, fox.WithIgnoreTrailingSlash(true))

// Register the route for the GET method
_ = f.HandleRoute(http.MethodGet, route)

// Update the route if needed
_ = f.UpdateRoute(http.MethodGet, route)
```

#### Deleting a Route

The API does not introduce a new method for deleting routes, as existing functionality suffices:

```go
_ = f.Delete(http.MethodGet, route.Pattern())
```

#### Registering New Endpoints for Existing Routes

One use case for this feature is to register a new HTTP method for all existing routes of a given method while maintaining their configuration.

Previously, achieving this required manually extracting route properties. Since routes do not expose middleware as intended, retrieving the middleware for a route was also not possible. Users must keep the state separately (e.g., in an annotation for the route).

```go
it := f.Iter()
for _, route := range it.Prefix(slices.Values([]string{"GET"}), "") {
    enable := route.IgnoreTrailingSlashEnabled()
    resolver := route.ClientIPResolver()
    
    _, _ = f.Handle(
        http.MethodPost,
        route.Pattern(),
        Handler,
        fox.WithIgnoreTrailingSlash(enable),
        fox.WithClientIPResolver(resolver),
        // Add some middleware from elsewhere
    )
}
```

With the new API, registering a new method for all existing routes is significantly simpler:

```go
for _, route := range it.Prefix(slices.Values([]string{"GET"}), "") {
    _ = f.HandleRoute(http.MethodPost, route)
}
```

Since route are strictly immutable once created, this new API make easier to swap route back and forth safely.

### Differential Benchmark
````
goos: linux
goarch: amd64
pkg: github.com/tigerwill90/fox
cpu: Intel(R) Core(TM) i9-9900K CPU @ 3.60GHz
                          │   old.txt   │              new.txt               │
                          │   sec/op    │   sec/op     vs base               │
StaticAll-8                 19.39µ ± 1%   19.44µ ± 0%  +0.21% (p=0.022 n=10)
GithubParamsAll-8           130.9n ± 1%   131.0n ± 1%       ~ (p=0.205 n=10)
InfixCatchAll-8             163.4n ± 1%   163.9n ± 0%  +0.28% (p=0.005 n=10)
LongParam-8                 73.04n ± 0%   73.05n ± 0%       ~ (p=0.493 n=10)
OverlappingRoute-8          154.6n ± 0%   155.6n ± 1%       ~ (p=0.469 n=10)
WithIgnoreTrailingSlash-8   118.7n ± 0%   119.0n ± 1%  +0.25% (p=0.005 n=10)
StaticParallel-8            20.84n ± 3%   21.31n ± 4%       ~ (p=0.218 n=10)
CatchAll-8                  71.45n ± 0%   72.63n ± 2%  +1.65% (p=0.001 n=10)
CatchAllParallel-8          10.35n ± 3%   10.24n ± 7%       ~ (p=0.540 n=10)
CloneWith-8                 155.9n ± 0%   157.9n ± 1%  +1.31% (p=0.000 n=10)
geomean                     129.6n        130.3n       +0.56%
````